Update index-en.html
Changed browser version number to text - version numbers are alphanumeric and have punctuation and other weird characters. 
Changed second date range question to refer to a different date range for a  'winter vacation'  rather than same as summer vacation, (so it's easier for the participant to refer to when thinking aloud) and changed country to refer to winter vacation country so it's a little less likely to be Canada. 

What I'd really like to do next is break this into a couple of pages, so it will be easier for me to time the tasks. But I don't know how to do that on Github, and I really want to do my first ever Fork and Pull Request, so let's just get these changes in for now.